voice_assistant/requirements.txt
Flask==2.3.3
SpeechRecognition==3.10.0
pyttsx3==2.90
gTTS==2.5.1
langdetect==1.0.9
google-generativeai==0.3.2